requirements.txt
bottle # webserver
pyyaml # to read config file
pandas
pyparsing
scikit-learn